I created a loop
Instead of reading the Excel workbook 9 times and extracting each sheet to create a data frame, I wrote a loop that reads through the workbook and creates a data frame out of each of the worksheets and assigns a name directly. this loop works well cos we don't have so many worksheets, there is a more efficient way that will work if our workbook was larger.